malloc =/= now

+++ source/dungeon/minimap_system/put_pixel.c
@@ -17,7 +17,7 @@ framebuffer_t *framebuffer_create(unsigned int width, unsigned int height)
 	buffer->width = width;
 	buffer->height = height;
 	buffer->pixels = malloc(width * height * (bitsPerPixel / 8));
-	for (int i = 0; i < (width * height * (bitsPerPixel / 8)); i++)
+	for (unsigned int i = 0; i < (width * height * (bitsPerPixel / 8)); i++)
 		buffer->pixels[i] = 255;
 	return (buffer);
 }
@@ -26,7 +26,8 @@ void my_put_pixel(framebuffer_t *buffer, int x, int y, sfColor color)
 {
 	int a = 0;
 
-	if (x < 0 || buffer->width <= x || y < 0 || buffer->height <= y)
+	if (x < 0 || buffer->width <= (unsigned)x ||
+		y < 0 || buffer->height <= (unsigned)y)
 		return;
 	a = (y * buffer->width + x) * 4;
 	buffer->pixels[a] = color.r;
